Create README.md
To Predict fraudulent credit card transactions with the help of machine learning models. Here we will analyse customer- level data that has been collected during the research collaboration of Worldwide and Machine Learning Group.

The dataset is taken from the Kaggle website. The data has a total of 284,807 transactions, out of which 492 are fraudulent. This signifies that the data is highly imbalanced. So, it is necessary to handle the imbalance before building the model.